DIRE-based-lab/requirements.txt
# DIRE-based Lab Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Diffusion models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Model zoo
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning utilities
scikit-learn>=1.2.0

# Configuration
PyYAML>=6.0

# Progress bar
tqdm>=4.65.0
